Revert deferrable refactorings
