+++ assets/js/components/InstanceOverview/InstanceOverview.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { getHost } from '@state/selectors';
-import { getCluster } from '@state/selectors/cluster';
+import { getCluster, getHost } from '@state/selectors';
 import HealthIcon from '@components/Health';
 import { Features } from '@components/SapSystemDetails';
 import { DATABASE_TYPE } from '@lib/model';
