+++ assets/js/state/actions/notifications.js
@@ -4,15 +4,3 @@ export const notify = ({ text, icon }) => ({
   type: NOTIFICATION,
   payload: { text, icon },
 });
-
-export const notifyChecksExecutionRequested = (clusterName) =>
-  notify({
-    text: `Checks execution requested, cluster: ${clusterName}`,
-    icon: '🐰',
-  });
-
-export const notifyChecksExecutionRequestFailed = (clusterName) =>
-  notify({
-    text: `Unable to start execution for cluster: ${clusterName}`,
-    icon: '❌',
-  });
