+++ assets/js/state/selectors/cluster.js
@@ -1,7 +1,4 @@
-export const getCluster =
-  (id) =>
-  ({ clustersList }) =>
-    clustersList.clusters.find((cluster) => cluster.id === id);
+import { getCluster } from '@state/selectors';
 
 export const getClusterHostIDs =
   (clusterID) =>
@@ -9,6 +6,9 @@ export const getClusterHostIDs =
     hostsList.hosts
       .filter((host) => host.cluster_id === clusterID)
       .map(({ id: hostID }) => hostID);
+
+export const getClusterSelectedChecks = (clusterID) => (state) =>
+  getCluster(clusterID)(state).selected_checks;
 
 export const getClusterName =
   (clusterID) =>
