+++ assets/js/state/selectors/cluster.test.js
@@ -1,4 +1,4 @@
-import { getClusterHostIDs } from './cluster';
+import { getClusterHostIDs, getClusterSelectedChecks } from './cluster';
 
 describe('Cluster selector', () => {
   it('should return the cluster hosts IDs', () => {
@@ -22,5 +22,31 @@ describe('Cluster selector', () => {
     };
 
     expect(getClusterHostIDs('cluster1')(state)).toEqual(['id1', 'id2']);
+  });
+
+  it('should return the cluster selected checks', () => {
+    const state = {
+      clustersList: {
+        clusters: [
+          {
+            id: 'cluster1',
+            selected_checks: ['check1', 'check2'],
+          },
+          {
+            id: 'cluster2',
+            selected_checks: ['check3', 'check4'],
+          },
+          {
+            id: 'cluster3',
+            selected_checks: ['check5', 'check6'],
+          },
+        ],
+      },
+    };
+
+    expect(getClusterSelectedChecks('cluster1')(state)).toEqual([
+      'check1',
+      'check2',
+    ]);
   });
 });
